# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for Codin-PIS. Authentication uses NextAuth session cookies (JWT). For Try it out: stay logged in in the same browser, or Authorize sessionCookie with authjs.session-token from DevTools."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://codin-pis-yetb.vercel.app",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Optional; APIs primarily use session cookies."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "NextAuth v5 session cookie (same host as the app)."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/rfqs/{id}": {
      "get": {
        "summary": "Get one RFQ with suppliers and line items",
        "description": "Requires rfqs.read. Returns requisition line items (from requisition_items) and invited suppliers.\nUser must belong to the RFQ's organization.\n",
        "tags": [
          "Procurement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RFQ detail in data envelope (suppliers, items, rfqNumber, status, etc.)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/rfqs/{id}/status": {
      "patch": {
        "summary": "Update RFQ workflow status",
        "description": "Requires rfqs.update. Valid status values — draft, sent, evaluating, awarded, cancelled.",
        "tags": [
          "Procurement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "sent",
                      "evaluating",
                      "awarded",
                      "cancelled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated (data.id, data.status)"
          },
          "400": {
            "description": "Missing or invalid status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "RFQ not found"
          }
        }
      }
    },
    "/api/rfqs": {
      "get": {
        "summary": "List RFQs for an organization",
        "description": "Requires rfqs.read. branchId must match the signed-in user's organization.",
        "tags": [
          "Procurement"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Organization UUID"
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "draft | sent | evaluating | awarded | cancelled"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Filter by title, rfqNumber, or category"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list (data, total, page, pageSize)"
          },
          "400": {
            "description": "Missing branchId"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create an RFQ from an approved requisition",
        "description": "Requires rfqs.create. requisitionId is required; the requisition must be **approved** and in the same organization.\nLine items are not posted — they are read from the requisition's requisition_items on GET.\nsupplierIds is optional (deduplicated server-side).\n",
        "tags": [
          "Procurement"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "branchId",
                  "requisitionId",
                  "title"
                ],
                "properties": {
                  "branchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "organizationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "requisitionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "paymentTerms": {
                    "type": "string"
                  },
                  "requiredDelivery": {
                    "type": "string",
                    "description": "ISO date e.g. 2026-08-01"
                  },
                  "deadline": {
                    "type": "string",
                    "description": "Supplier response deadline (ISO or display string)"
                  },
                  "supplierIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created (data includes suppliers and items from requisition)"
          },
          "400": {
            "description": "Validation error (e.g. requisition not approved or wrong org)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
...
}